+++ src/frontend/components/Header.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 
-
 function Header() {
     return (
         <header>
